bg/lwp-619_alpa_beta : fix review comments
