+++ src/include/miopen/conv/problem_description.hpp
@@ -133,6 +133,22 @@ constexpr TElement GetW5(unsigned spatial_dims, const std::vector<TElement>& dat
 
 namespace conv {
 
+enum class AlphaBetaCase
+{
+    /* IDENTITY      alpha = 1.0 and beta = 0.0 */
+    /* SCALE         alpha = 4.2 and beta = 0.0 */
+    /* BILINEAR      alpha = 3.2 and beta = 1.1 */
+    /* ERROR_STATE   alpha = 0.0 and beta = 3.1 */
+
+    IDENTITY    = 0, /* alpha = 1.0 and beta = 0.0.*/
+    SCALE       = 1, /* alpha with some value and beta 0.0*/
+    BILINEAR    = 2, /* both alpha and beta with some value*/
+    ERROR_STATE = 3, /* alpha 0.0 and beta with some value, this should not occur.
+                        But used to check for errors.*/
+};
+
+AlphaBetaCase GetAlphaBetaCase(const Scalar& alpha, const Scalar& beta);
+
 struct ProblemDescription : ProblemDescriptionBase
 #if MIOPEN_ENABLE_SQLITE
     ,
@@ -148,8 +164,8 @@ struct ProblemDescription : ProblemDescriptionBase
                        const ConvolutionDescriptor& conv_,
                        Direction direction_,
                        int bias_            = 0,
-                       const Scalar& alpha_ = {1.0},
-                       const Scalar& beta_  = {0.0})
+                       const Scalar& alpha_ = Scalar(1.0),
+                       const Scalar& beta_  = Scalar(0.0))
         : in(in_),
           weights(weights_),
           out(out_),
@@ -160,10 +176,10 @@ struct ProblemDescription : ProblemDescriptionBase
           direction(direction_),
           bias(bias_),
           alpha(alpha_),
-          beta(beta_)
+          beta(beta_),
+          alpha_beta_case(GetAlphaBetaCase(alpha, beta))
     {
         HeuristicUpdateLayouts();
-        alpha_beta_case = GetAlphaBetaCase(alpha, beta);
     }
 
     // Conv descriptor getters
@@ -484,8 +500,8 @@ private:
     std::string out_layout;
     Direction direction           = Direction::Forward;
     int bias                      = 0;
-    Scalar alpha                  = {1.0};
-    Scalar beta                   = {0.0};
+    Scalar alpha                  = Scalar(1.0);
+    Scalar beta                   = Scalar(0.0);
     AlphaBetaCase alpha_beta_case = AlphaBetaCase::IDENTITY;
 };
