+++ test/gpu_reference_kernel.cpp
@@ -422,7 +422,7 @@ struct gpu_reference_conv_2d : gpu_reference_kernel_base
                 rand_tensor_integer(in);
                 rand_tensor_integer(wei);
                 /// \ref copy_non_packed_output_before_convolution
-                // rand_tensor_integer(out);
+                rand_tensor_integer(out);
 #if MIOPEN_BACKEND_OPENCL
                 status = clEnqueueWriteBuffer(q,
                                               in_dev,
@@ -502,7 +502,7 @@ struct gpu_reference_conv_2d : gpu_reference_kernel_base
                 rand_tensor_integer(out);
                 rand_tensor_integer(wei);
                 /// \ref copy_non_packed_output_before_convolution
-                // rand_tensor_integer(in);
+                rand_tensor_integer(in);
 #if MIOPEN_BACKEND_OPENCL
                 status = clEnqueueWriteBuffer(q,
                                               out_dev,
@@ -575,7 +575,7 @@ struct gpu_reference_conv_2d : gpu_reference_kernel_base
                 rand_tensor_integer(in);
                 rand_tensor_integer(out);
                 /// \ref copy_non_packed_output_before_convolution
-                // rand_tensor_integer(wei);
+                rand_tensor_integer(wei);
 #if MIOPEN_BACKEND_OPENCL
                 status |= clEnqueueWriteBuffer(q,
                                                in_dev,
@@ -791,7 +791,7 @@ struct gpu_reference_conv_3d : gpu_reference_kernel_base
                 rand_tensor_integer(in);
                 rand_tensor_integer(wei);
                 /// \ref copy_non_packed_output_before_convolution
-                // rand_tensor_integer(out);
+                rand_tensor_integer(out);
 #if MIOPEN_BACKEND_OPENCL
                 status = clEnqueueWriteBuffer(q,
                                               in_dev,
@@ -865,7 +865,7 @@ struct gpu_reference_conv_3d : gpu_reference_kernel_base
                 rand_tensor_integer(out);
                 rand_tensor_integer(wei);
                 /// \ref copy_non_packed_output_before_convolution
-                // rand_tensor_integer(in);
+                rand_tensor_integer(in);
 #if MIOPEN_BACKEND_OPENCL
                 status = clEnqueueWriteBuffer(q,
                                               out_dev,
